// src/api/public-ws/hyperliquid.models.ts
/**
 * Таймфреймы для Hyperliquid API
 * Поддерживаемые значения: 1m, 3m, 5m, 15m, 30m, 1h, 2h, 4h, 8h, 12h, 1d, 3d, 1w, 1M
 */
export enum HyperliquidTimeframe {
  Min1 = '1m',
  Min3 = '3m',
  Min5 = '5m',
  Min15 = '15m',
  Min30 = '30m',
  Hour1 = '1h',
  Hour2 = '2h',
  Hour4 = '4h',
  Hour8 = '8h',
  Hour12 = '12h',
  Day = '1d',
  Day3 = '3d',
  Week = '1w',
  Month = '1M',
}